etl manifest: add `samples-haarvi` to accepted collections
These are barcodes for samples aliquoted by Chu lab that are then sent
to BBI for testing. Even though this is technically a "sample" barcode,
it is effectively acting as a "collection" barcode for BBI.

BBI does assign an actual sample barcode to these as well, so accepting
the `samples-haarvi` barcode as a collection identifier does not break
anything.