Add `collected` to `warehouse.sample`
In the first year of the SFS study, sample collection date and
encounter date were the same for in person swabs at kiosks, so we did
not need a separate column for collection date.

This is no longer true in Year 2 and in the SCAN project. The encounter
date and collection date can differ because participants first fill out
the questionnaire then receive a swab kit on a later date.

There is no reliable `collection_date` field across all REDCap projects,
so we will be relying on the lab's aliquoting manifest to fill in this
column.