redcap: Add a Record class (a subclass a dict)
Using a custom subclass of dict lets us provide some custom attributes
like .project to point to the record's Project instance and .id to grab
the record's primary unique id (regardless of what it is named).

This is helpful when programmatically processing multiple projects that
share the same structure, for example with SCAN's project-per-language
setup.