etl fhir: Update assert_required_resource_types_present
Update the requirement for Observation Resources to allow multiple
Encounters in the same bundle.

Since the Specimen Resource will only be linked to one of the
Encounters, the number of Observations will not be greater than the
number of Encounters. Therefore, update to require the number of
Observations to be greater than or equal to the number of Specimens
instead.