redcap: Support fetching the REDCap version
This is useful/necessary when constructing links into REDCap if the
instance embeds the current version into the URLs, as both ITHS and FH
do.  While URLs with previous versions will often (but not always)
redirect to URLs for the latest version, an actual previous version is
still needed; there's no versionless URLs.