Add materialized-view-refresher role
Regular users cannot refresh materialized views because only the view
owner has permission to refresh them. Create a function that runs in
the security context of its owner so that other users can refresh the
view. Grant execute on function to "materialized-view-refresher" so that
all users with that role can refresh materializeed views.