Make manifest parsing support multiple source types
In order to support new manifests (to supplement the current aliquoting
manifest), this commit adds a collection_column setting and a
collected_date_column setting (and
corresponding JSON keys) so that now each of the 'identifier',
'collection_identifier' and 'collected' columns in the
warehouse.sample table that get filled
by the manifest ETL have a defined setting to explicitly indicate
which source column goes in to which table column. The sample,
collection, and collected_date columns are deemed "reserved", and
a check ensures that a column with the same name is not added to the
extra_columns set. This change will require the alliquot manifest
YAML files be edited slightly. That is in PR
https://github.com/seattleflu/specimen-manifests/pull/15 .

In the manifest ETL module a new step is added to identify which
of the "reserved" keys appear in the data, and validation and
update logic uses those flags. For example, don't update the
collected column unless the source data has the collected_date key in
its JSON. (This key is written only if the manifest's YAML maps a
column as collected_date_column.)