etl fhir: Handle more edge cases during Encounter.meta.source handling
Use core libraries for parsing data: URIs to support all variants.

Warn about exceptions and debug log detailed information, but continue
processing by returning the source URI as-is.  The idea is that this
handling is best-effort and not critical to ingesting the rest of the
Encounter or larger FHIR bundle.