Use etag for digest for incident report manifest
Currently, the incident report manifest parser uses a hash
of the entire Google Sheets document. We've found that the contents
of some metadata change even when the data content of the spreadsheet
itself doesn't change. This leads to a different digest even when
the speadsheet hasn't changed.

This commit switches to using the etag of the Google document
for the digest.