@with_database_session: Optionally pass the selected database action to commands
The decorator still performs the specified action, but this gives
decorated commands a way to know what mode they're running under.  This
is useful, for example, if the command accesses non-database resources
and wants to extend dry run mode to them as well, as is the case with a
new command in our local customizations which will modify data in
REDCap.

The API of the decorator remains the same unless commands use the new
option, so existing callers do not need to be updated.