etl manifest: Change missing identifiers warning
Instead of checking if there are no sample or collection identifiers
associated with a sample in the manifest, check instead for sample and
collection barcodes. This will handle the situation where
`find_identifier` (which does its own logging if it will return None)
never gets called.

Keep the log level at "warning" because if we're ingesting manifest
records with neither sample nor collection barcodes, it means something
is awry in our manifest parse/upload process. We would expect to drop
rows with neither barcode available.