@with_database_session: Use an enum for CLI option flag values
Not necessarily warranted for a small, self-contained bit of code where
the value initialization is immediately adjacent to the only value
comparisons, but I'll soon add the ability to pass the action flag value
to the command itself.  The command will need a way to safely compare
values to known action modes.