@with_database_session: Internally refactor to use dynamic decorator generation
The final decorated command (i.e. the return value) is now produced by
calling an internal decorator function.  This extra level of indirection
allows the decoration to dynamically respond to options instead of being
statically defined.  This is in preparation for adding an option to
@with_database_session.

No behavioural changes, mostly indentation-only changes.