etl redcap-det: Accept env var for geocoding cache
Instead of having to specify a geocoding cache file on each invocation
of a REDCap DET ETL, accept an environment variable that will map to the
correct file. This will reduce the verbosity of invoking multiple REDCap
DETs that share a cache, for example, in a list of scheduled cron jobs.