Optimize barcode distance check by only finding first conflict
While during initial development it was useful to know what the full set
of conflicting barcodes was, it's wholly unnecessary work during
production.  The first conflict found is sufficient to fail the insert
with an exclusion violation.

In my testing (3 trials of 100 barcodes on my local computer), this
reduces the mean time per identifier from 2.24s to 2.19s.

I tried a variant of this change using exists(…) instead of limit 1, but
it performed the same.  I also tried replacing the "except select
old_barcode" statement with two where-clause variants:

    and (old_barcode is null or old_barcode != barcode)

    and barcode is distinct from old_barcode

but both performed worse.  I believe this is because Pg can parallelize
the query when the Hamming distance is the only where-clause criteria.