Add a short-circuiting version of our Hamming distance function
This version reports exact distances for values less than or equal
(LTE) the given threshold and a distance of at least threshold + 1 for
anything greater.  That is, it stops computing distance once the
distance exceeds the threshold.  This short-circuiting is inspired by
the levenshtein_less_equal() function from fuzzystrmatch.

In a subsequent commit, I'll be swapping to this new function in the
warehouse.identifier distance check trigger.