Optimize barcode distance check by switching to hamming_distance_lte()
This version of the Hamming distance function will short-circuit once
it's clear that the barcode in question is not in conflict, thus
reducing the number of per-position string comparisons.

In my testing (3 trials of 100 barcodes on my local computer), this
reduces the mean time per identifier from 1.91s to 1.25s.