Index receiving.redcap_det.document::jsonb
The REDCap DET ETLs (run as `id3c etl redcap-det …`) query on the DET
document casted to jsonb (from json) in order to use the containment
operator, @>.  When this started becoming slow at the beginning of the
pandemic, I created a rogue index outside of Sqitch¹ due to extreme time
pressure, intending to later followup with a more systematic change to
the document data type from json → jsonb.  This is me finally following
up, after 238 days, by re-creating the same index as a properly managed
Sqitch change.

The larger, more systematic change I assumed we'd have time to consider
and implement is still up for consideration, but doesn't have to block
bringing this rogue index into the fold.

The index isn't too large that we need to avoid re-creating it, so when
this change is deployed to create the same index under its "blessed"
name, I'll drop the rogue one.

¹ receiving.redcap_det_document_jsonb_idx__trs_20200323