Add barcode minting optimizations
Create a new column `slices` on the warehouse.identifier table
to hold 2-character-wide slices of the barcode text. These values are
used to find similar barcodes before running the hamming distance
function.

The changes are:
1. Add the `barcode_slices` function to create the slices array
from the input barcode.

2. Add the `slices` column to the warehouse.identifier table. Update
existing rows to populate `slices` using the `barcodes_slices` function.

3. Update the barcode-distance-check trigger function that is
used when warehouse.identifier rows are created or updated. Add this line
to leverage the slices column before calling the hamming_distance_lte
function: `where slices && NEW.slices`
That is, only call `hamming_distance_lte` on rows that have barcode slices
that intersect with the candidate slices. Previously, `hamming_distance_lte`
was called on all the barcodes in the table.

A compromise I made was to use this same trigger function to populate a
new or updated row's `slices` column as well as to use it to look for
conflicts. I did this to avoid adding the overhead of a new trigger
function that would only set this value. Philosophically it's odd
for a function with "check" in its name to change data in the row being
checked. I hesitated to change the name given our desire to keep names
short.

I couldn't add the `slices` population to the `barcode-default-from-uuid`
function because it is used only for inserts into the warehouse.identifier
table. I needed to ensure that `slices` get updated when rows get updated
as well.

See the schema/verify/functions/barcode_slices.sql file for examples
of what the slices look like.

Minting barcodes is about 4x faster (across multiple tests) with these
changes applied. In the most recent test of minting 1275
barcodes here were the times and the number of conflicting barcodes
encountered. Note that the number of conflicts is about the same
between the runs, indicating that the conflict detection logic is
comparable.

Before changes:
Elapsed time: 59 minutes, Minted 1275 identifiers in 2285 tries (1010 retries)

With changes:
Elapsed time: 15 minutes, Minted 1275 identifiers in 2444 tries (1169 retries)