etl presence-absence: skip samples that have been marked as failed
Each sample has a top level key "sampleFailed" that is marked True if
the entire sample has been marked as failed. We need to inspect this
top level key because the individual target results do not reflect this
failure status.

ID3C currently does not model failed presence/absence results,
so skip all results if the sample is failed at this top level.