Revert "api: Explicitly close DatabaseSession at request end"
This reverts commit 8761572316a4b08b2740a58990571999d00fadcd.

Explicitly closing the database connection in the route decorator is too
early for streaming responses because the response data generator holds
a reference to the connection (and uses it) after the route function
returns.

This was causing errors for our custom API routes that provide data out:

    psycopg2.InterfaceError: connection already closed

and was first noticed with the endpoints used by the SCAN dashboard.

There is likely a solution that preserves the explicit close by
registering a Flask context teardown/after request callback, but punt on
that additional complexity for now.  The behaviour this returns the API
to was seemingly not problematic in practice, at least in production.