etl redcap-det: Use new-style environment variables for REDCap API credentials
The new-style is the future and makes it easier to organize credentials,
switch between projects, and programmatically invoke this command.

Notably, the same REDCap instance URL (provided by the ETL routine) is
used to both select from receiving.redcap_det and find the right
credentials.  Thus, a URL-matching safety check is no longer necessary
because of this better design.