Remove retry limit for barcode minting
This commit removes the logic in the `mint_identifiers` function that throws
an error if a maximum number of minting retries are hit. When we hit that
currently, we simply re-run the job. With this maximum retry limit removed,
the job will just keep retrying until it gets a valid barcode.