Rework barcode_slices function to resolve db restore issue
Since the introduction of the barcode slices functionality, restoring
the production database has been failing with errors that look like:
`pg_restore: [archiver (db)] Error from TOC entry 6314; 1259 4123860
INDEX identifier_barcode_slices_idx postgres
pg_restore: [archiver (db)] could not execute query: ERROR:
type "citext[]" does not exist
LINE 5:         slices citext[];'

The following link showed a similar issue, and the author figured out that
when pg_restore creates functional indexes on tables in a schema different
from the schema in which the function is defined, that it fails
to find the function. His solution is to explictly prefix the function's
schema when it is used in the indexes.
https://www.postgresonline.com/journal/archives/121-Restore-of-functional-indexes-gotcha.html

Taking that as a clue, the solution for us is to explictly specify the `public`
schema in the function when we refer to the `ci_text` extension so that
when pg_restore restores the indexes that use the function, the
`ci_text` extension will be found.

In the sqitch plan I added a tag that I forgot to add as part of the
barcodes slices deployment a couple of months ago.