json: Add load_json() function which contextualizes decoding errors
Snips the source document around the error position for context and
marks the problematic character.

One slight concern here is the chance for sensitive information to slip
in the logs via JSON content (e.g. received from the REDCap API).  I
think this is likely to be ok, as the expectation is not that malformed
JSON is sent, but that non-JSON text is are sent.  In the case of
malformed JSON, the document will be snipped to 21 chars, reducing both
log message length and chance of sensitive information leaking.