identifier: restrict `copies-per-barcode` option
Make the `--copies-per-barcode` option a click.Choice type restricted to
the values 1 or 2. This is the easiest way to avoid bad UX for the
label PDFs where a single barcode can span multiple rows.

In the future we may consider using the label SKU type to calculate the
blanks needed to prevent the same barcode from spanning multiple rows.