Migrate tests from Travis CI to GitHub Actions
We ran out of free trial credits on Travis CI back in January.  Rather
than apply for a one-time grant of open-source credits (and have to
re-apply in the future when they run out again, whenever that might be),
move to GitHub Actions where this isn't an issue.  We're already using
GitHub for repository hosting and PRs, so using it for CI too makes
sense and reduces our service sprawl.

I based this CI workflow on the one I setup and honed for the Nextstrain
CLI repo a month or two ago.  It supports a test matrix of OS and Python
versions.  Only Ubuntu is used for now, since psycopg2 has build issues
on GitHub's macOS runners.

Currently only Python 3.6 is tested since that's what our Pipenv
environment requires.  (The Travis config ran a job with 3.7, but then
the tests in that job ran on 3.6.  This was an oversight.)  A subsequent
commit will add more Python versions after some related Pipenv/setup.py
shuffling.