etl presence-absence: Fix type hint mistake
We pass a datetime.date object to this function, instantiated by the
database layer when we fetch the row, and then within this function
compare the received_date to a manually instantiated date object.  MyPy
caught this mistake, although it's an error in hinting only.

Tests now pass for me.