Move development deps out of Pipfile and into setup.py as the "dev" extra
These were originally in Pipfile because we use Pipenv as a common
development environment.  In order to test multiple Python versions in
CI, however, it's easier to track the dev deps in setup.py and install
them from outside Pipenv.  The Pipenv dev env remains functionally the
same, with the bonus of not needing to remember the --dev option to
`pipenv install` and `pipenv sync`.