api: Add endpoints for managing identifier sets
I wrote these endpoints over 2 years ago while integrating identifier
minting and barcode label generation, only to end up refactoring the
bulk of it out and generating labels via a CLI command instead.  With a
web interface for minting new identifiers in mind, I kept the code
around as a wip branch.  I thought of it again as we're thinking about
identifier API integration with the BBI LIMS, so I updated the code to
account for changes that happened in the interim.