Update manifest parse_using_config() to allow lists of workbooks
While continuing to support one configuration entry per workbook, this change will allow
workbooks that share the same format to be entered as a list within a single entry. This
prevents needing to copy and paste the whole config (for example, when a new aliquoting
sheet is added), reducing the potential for errors and resulting in smaller YAML files.