Add barcode-uses/verify endpoint to API
Adding /verification/barcode-use/verify endpoint to API for use with Laboratory Information
Management System (LIMS) and/or other system integration, for automation of QC processes.

This endpoint takes a JSON array of objects as input, with each object containing a barcode
and its associated use type to verify (i.e. `sample`, `collection`, `clia`). The response
will be a JSON array of the same objects with boolean values for `barcode_found` to indicate
whether the barcode exists in ID3C and `use_match` to indicate whether the input use type
matches what is stored in ID3C.

Updating API documentation page.