Updates to API to include new `use` field where relevant.
The `identifier_set.use` column is useful as an additional data point in the
`identifer` and `identifier-set` GET endpoints, and is a required input for
creating new identifier sets through the `identifier-sets` PUT endpoint.

A new GET endpoint useful for listing all values in the new `identifier_set_use`
table.