Retry failed REDCap API calls
REDCap API has a bug that incorrectly returns 200 status code and HTML
response text in HTML format containing an "unknown error" and "This may
have occurred if you have opened multiple browser tabs of the same REDCap
page". Requests that receive this reponse often succeed with additional
attempts.

This change will search for that specific response and retry
the request up to 10 times before proceeding. If the limit of 10 retries
is reached and the same response is still being recieved, the function
will continue as it did before with the invalid response text. That will
result in the same exception that was occurring previously, which happens
when the response text is being processed within other parts of the code
that call this function. This error handling may be improved in the future,
but that behavior is being left in place to ensure alerts are still surfaced
in the same manner for now.