Add request body json validation
Adds jsonschema package to validate format of request body. A valid
schema is defined and then used to validate the incoming JSON before
continuing with verification of the barcodes and use types. If the
incoming JSON does not match the defined schema, a 400 (Bad Request)
status code is returned along with jsonschema's ValidationError text
which contains detailed message on why the validation failed.

Similar validation can (and probably should) be implemented on other
API endpoints going forward, but for now is just being added to the
endpoint specific to this branch.