Install PyCap package as alternative to id3c.redcap
PyCap is an interface to the REDCap Application Programming Interface (API)
developed by Vanderbilt University and Childrens Hospital of Philadelphia. It
is an alternative to id3c's redcap library which has, in early testing, had
better performance and results for retrieving large numbers of REDCap records.

Initial use case being targeted is the backoffice `uw-ehs-report/export-redcap-uw-reopening`
script, but there may be many others where this can be used in place of id3c.redcap for
pulling REDCap records.