Sanitize SmartyStreets address inputs
SmartyStreets requires that the input to its address-geocoding
API be a Latin-1-encodeable string. Since it's specifically
checking that its input is a string, we can't send it bytes,
which would make this easy.

Instead, sanitize the input, replacing Unicode codepoints that can't
be encoded in Latin-1 with a transliterated string representation
that can, or a space by default.

This should be considered a first-pass attempt.