Adds more control over how collection date is assigned.
Update to upsert_sample function which is used in multiple contexts
to control whether or not the incoming collection_date will be used to
overwrite an existing value in the `collected` column of the sample record.

For samples updated through the API, which are coming from the lab,
the collection date will overwrite the existing value. For records coming
from FHIR ETL which originate from REDCap, the collection_date will only
be used if there is no value in the table.

The manifest and enrollments ETL are no longer in production use, but are
updated here to keep them functional.