Updating GET sample parameters to differentiate barcode types
Rather than accpting either sample or collection barcodes as a
positional argument, a collection_barcode URL parameter is
now required to search on collection barcode. This makes use of
the endpoint more clear and intentional which regard to which
type of barcode is being used to retrieve the sample record.