Quieting alerts for HCT unrecognized barcodes
The volume of unrecognized barcodes for HCT project has increased
since start of dropbox collection. The QC process for resolving these
in REDCap is being implemented in LIMS due to the volume and need for
a UI to manage them. The alerts related to HCT unrecognized barcodes
are being quieted here in favor of the new feature in LIMS.