Prompt user when geocoding in non-production environment
Add a prompt when requests to SmartyStreets occur in a non-production
environment. This is to prevent accidental usage of credits while
testing processes in development. An input prompt will appear prior
to the request to SmartyStreets with options yes/no/all/none.

During local development it is typically fine to skip geocoding any new
addresses.