Make geocoding cache file required
Make geocoding cache filename a required option for the REDCAP DET ETL command, and
check to make sure the file exists. This is being changed to prevent accidental
usage of Smartystreets credits. The cache file should always be used, and
downloaded from production prior to running any REDCAP DET ETL command locally.

The previous behavior did not require a cache file to be specified at all, and would
automatically create a new one if the one specified did not exist. This made it too
easy to use SmartyStreets credits locally without realizing it.