Add option to create the cache file if it does not exist
This option is being added back into to allow doctests to continue
to work after making the cache file required. Pytest was failing because
there is no existing cache file when running tests.

The default will still be to not create a new cache file if the specified
filename does not exist, but we will have the option to be able
to create the file for things like automated testing.