Adding Github action to automatically publish to pypi
Publishing id3c to pypi will help standardize how this package is being
installed compared to other dependencies. This will make the package simpler
include in multiple applications and pin to a specific version when needed.

One workflow change this will require is that commits on master branch will need to
be tagged to be published to pypi.org (whereas all commits on master branch will be
published to test.pypi.org). A version naming convention will need to be
decided on and integrated into the deployment workflow, but the specific format
chosen will not impact this PR.