Remove publishing to test.pypi.org
Publishing to test.pypi.org on each push to master creates a couple of
challenges. First pypi cannot overwrite a file with the same name,
which is the case when publishing a package that does not have a different
reference tag in github. There are some solutions to work around this by
having a specific "RC" tag format for release candidates and only publishing
those to test.pypi.org. At this point that would create additional unnecessary
steps in the deployment process but may be worth revisiting in the future once
there is a versioning/tagging conventions are established.