Add timeout to REDCap requests
A 300 second timeout is being added to all REDCap requests, which ultimately
all use the _fetch function. It has been confirmed there is 120 second timeout on the
server side so this should be plenty of extra time before giving up on a request.

The goal of this change is to prevent requests from hanging indefinitely, and to
throw an exception when a request exceeds 5 minutes.