Update FHIR ETL to include Immunization entries
New entries are being added to the FHIR bundle for Immunization as part
of the updated UW Retrospectives ETL seattleflu/id3c-customizations#267.

These entries are using standard FHIR format and coding, and are linked to
a FHIR Patient, so in ID3C they will be ingested into the `individual.details`
column rather than `encounter.details`.

The `individual.details` jsonb column has not been utilized yet in ID3C, so this
will be the only data present until other ETLs are updated to include Immunization
entries or other resource entries related to Patients are are added to the FHIR
bundle and this ETL.