Create new tiny duplicates base class
This change adds a new base `LCRY2380DuplicateLayout` class which
the `SamplesLayout` and new `CollectionsCascadiaTinySwabsHome`
classes inherit from. This class can serve as the new
base class which any collections that use tiny swabs in duplicate
can inherit from so that they can take advantage of the `blanks_before`
function.