Add aliquot C and rack C fields to sample JSON schema
To be able to accept aliquot and rack C values in sample POST requests,
the JSON schema is being updated to include the `aliquot_c`, `rack_c` and
`rack_c_nickname` attributes. These will then be included in arrays with
the corresponding A and B values to be stored in `sample.details`.