Refactor upsert_encounter to reduce gaps in sequence
The upsert_encounter function is using `on conflict` option to execute
an insert or update on each record. This results in the sequence being
incremented on each upsert, regardless of whether a record was inserted or
not, leaving large gaps in the sequence. This change will prevent that by first
checking if the record exists, then running an update or insert as appropriate.