Prevent updating of unchanged encounter records
The upsert_encounter method is unnecessarily updating records that have not
changed as part of our normal ETL processes. This results in bloat in the encounters
table as a large number of unchanged records are constantly being updated, creating
dead tuples that impact performance.

This change prevents updating records that have no changes in the relevant columns
by comparing a hash of the existing vs. the incoming values. If the hash values are
identical, the update will be skipped.