Move hash comparison to sql
To consolidate the code and improve performance, the hash calculation on the incoming
values is being moved to the initial `select` query, as well as the comparison of the
two hash values to determine if the data has changed. This should perform better and make
this code easier to maintain if there are future schema changes that require changes to
the `upsert_encounter` function.