Ignore FHIR resource reference IDs when determining if data changed
The current ETL processes for creating FHIR bundles include the generation
of random UUIDs which are used as reference identifiers within each bundle.
Since these values are not indicitive of the actual data changing, they
are being removed before comparing the existing and incoming data.

Also being removed is the md5 hashing, which was not necessary for the comparison.