Reduce Unnecessary Updates to `encounter_locations`
Similar to what was just fixed by Dave in the `warehouse.encounter` upsert function, this change
compares old values in the `warehouse.encounter_location` table with new incoming values from
the `receiving.fhir` table. If the values are the same, then this update is skipped, thereby
reducing bloat in this table.