Cast rows to text for data comparison in upsert_encounter
When comparing two items with data type `row`, the presence of a null element
within either row will return null rather than the expected boolean. This was
resulting in old encounters continuing to be updated even though the data was not changing.

To avoid this behavior, the rows are being cast to text before comparing them.