Add optional argument to dump_ndjson for masking columns
Adding an new argument to the dump_ndjson function to mask columns that
may contain sensitive data during development and testing. If no value is
passed, the default will be to keep all columns unmasked.

The `mask_values` function is being added here for this purpose, but will also
serve as a general function to be used in other contexts.