Refactor upsert encounter to avoid sequence gaps and unneccessary updates
In line with other upsert functions that were recently updated, this change
will prevent gaps in the individual_id sequence that are caused by the use of
`insert ... on conflict do update`. The function will now run a select query,
followed by an insert or update as appropriate. A data comparison is being
added to the select query to determine if the data is changing, which is then
used to avoid unnecessary updates that contribute to table bloat.